Model script is now split into functions.
