+++ model/test_functions.py
@@ -1,0 +1,54 @@
+import numpy as np
+import matplotlib.pyplot as plt
+import itertools
+from PIL import Image
+
+def plot_sample(input_array, index, title=''):
+    '''
+    Function to plot 28x28 pixel drawings that are stored in a numpy array.
+    Specify how many rows and cols of pictures to display (default 4x5).  
+    If the array contains less images than subplots selected, surplus subplots remain empty.
+    '''
+    
+    imgplot = plt.imshow(input_array[index,:784].reshape((28,28)), cmap='gray_r', interpolation='nearest')
+    plt.xticks([])
+    plt.yticks([])
+
+    plt.show()
+
+def plot_confusion_matrix(cm, classes,
+                          normalize=False,
+                          title='Confusion matrix',
+                          cmap=plt.cm.Blues):
+    """
+    This function prints and plots the confusion matrix.
+    Normalization can be applied by setting `normalize=True`.
+    """
+    plt.imshow(cm, interpolation='nearest', cmap=cmap)
+    plt.title(title)
+    plt.colorbar()
+    tick_marks = np.arange(len(classes))
+    plt.xticks(tick_marks, classes, rotation=45)
+    plt.yticks(tick_marks, classes)
+
+    if normalize:
+        cm = np.round(cm.astype('float') / cm.sum(axis=1)[:, np.newaxis], 5)
+        print("Normalized confusion matrix")
+    else:
+        print('Confusion matrix, without normalization')
+
+    print(cm)
+
+    thresh = cm.max() / 2.
+    for i, j in itertools.product(range(cm.shape[0]), range(cm.shape[1])):
+        plt.text(j, i, cm[i, j],
+                 horizontalalignment="center",
+                 color="white" if cm[i, j] > thresh else "black")
+    plt.tight_layout()
+    plt.ylabel('True label')
+    plt.xlabel('Predicted label')
+
+if __name__ == '__main__':
+    pic = Image.open('./axe.png')
+    axe = np.array(pic)/255
+    print(axe)
